# docs/api/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "forzion.tech.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/internal/processar-renovacoes-treinador": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "Processa renovações mensais de planos de treinadores (requer X-Internal-Key)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/internal/processar-renovacoes": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "Processa renovações mensais de assinaturas (requer X-Internal-Key)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/internal/reconciliar-pagamentos": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "Reconcilia eventos Stripe da janela especificada (default últimos 7d) — requer X-Internal-Key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ReconciliarPagamentosStripeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliarPagamentosStripeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/aluno/pagamentos/{pagamentoId}": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Obtém status e QR code de um pagamento",
        "parameters": [
          {
            "name": "pagamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagamentoResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/aluno/pagamentos/assinatura/{assinaturaId}": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Lista histórico de pagamentos de uma assinatura (paginado)",
        "parameters": [
          {
            "name": "assinaturaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListarPagamentosAssinaturaAlunoResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/treinador/pagamentos/cobrar/{assinaturaId}": {
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Gera cobrança mensal para uma assinatura (metodo: Pix ou Cartao)",
        "parameters": [
          {
            "name": "assinaturaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "metodo",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MetodoPagamento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagamentoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/treinador/pagamentos/recebimentos": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Lista recebimentos do treinador (paginação keyset por cursor)",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tamanho",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListarRecebimentosTreinadorResultado"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/treinador/plano/assinatura": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Retorna a assinatura de plano ativa do treinador, ou null se não houver",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/treinador/plano/pagamento/{pagamentoId}": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Retorna status de pagamento de plano do treinador",
        "parameters": [
          {
            "name": "pagamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/treinador/plano/trocar": {
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Troca o plano do treinador (upgrade/downgrade/regularização)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocarPlanoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrocarPlanoTreinadorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/treinador/plano/cobrar": {
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Gera cobrança de renovação do plano do treinador (metodo: Pix ou Cartao)",
        "parameters": [
          {
            "name": "metodo",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MetodoPagamento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IniciarPagamentoPlanoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ListarPagamentosAssinaturaAlunoResponse": {
        "required": [
          "items",
          "total",
          "pagina",
          "tamanhoPagina"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PagamentoResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tamanhoPagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "MetodoPagamento": {
        "enum": [
          "Pix",
          "Cartao"
        ]
      },
...
      "PagamentoResponse": {
        "required": [
          "pagamentoId",
          "assinaturaAlunoId",
          "valor",
          "status",
          "metodoPagamento",
          "pixQrCode",
          "pixQrCodeUrl",
          "pixExpiracao",
          "clientSecret",
          "dataPagamento",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "pagamentoId": {
            "type": "string",
            "format": "uuid"
          },
          "assinaturaAlunoId": {
            "type": "string",
            "format": "uuid"
          },
          "valor": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/PagamentoStatus"
          },
          "metodoPagamento": {
            "$ref": "#/components/schemas/MetodoPagamento"
          },
          "pixQrCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "pixQrCodeUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "pixExpiracao": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "clientSecret": {
            "type": [
              "null",
              "string"
            ]
          },
          "dataPagamento": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReconciliarPagamentosStripeRequest": {
        "required": [
          "desdeUtc"
        ],
        "type": "object",
        "properties": {
          "desdeUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "ReconciliarPagamentosStripeResponse": {
        "required": [
          "totalEventos",
          "replayed",
          "jaConsistentes",
          "erros",
          "desdeUtc"
        ],
        "type": "object",
        "properties": {
          "totalEventos": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "replayed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "jaConsistentes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "erros": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "desdeUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TrocarPlanoRequest": {
        "required": [
          "planoPlataformaId"
        ],
        "type": "object",
        "properties": {
          "planoPlataformaId": {
            "type": "string",
            "format": "uuid"
          },
          "metodo": {
            "$ref": "#/components/schemas/MetodoPagamento"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Cole o JWT gerado pelo endpoint POST /auth/login aqui.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}